testsuite: move modctl proto tests to subdirectory
